backend/pyproject.toml
[tool.poetry]
name = "digital_doctors_helper"
version = "0.1.0"
description = "Backend for app ddh"
authors = ["Солнышко"]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
gunicorn = "20.1.0"
psycopg2-binary = "2.9.6"
django-health-check = "^3.17.0"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-cors-headers = "^3.14.0"
django-filter = "^23.2"
sentry-sdk = {extras = ["django"], version = "^1.21.1"}
django-cleanup = "^7.0.0"
celery = "^5.2.7"
redis = "^4.5.5"
flower = "^1.2.0"
openpyxl = "^3.1.2"
thefuzz = "^0.19.0"
python-levenshtein = "^0.21.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

